mind-guide/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-jose==3.3.0
bcrypt==4.2.1
pydantic<2
python-multipart==0.0.9
email-validator<2.1
groq>=1.1.0
